Fix build errors: add uint32_t casts to resolve type mismatch in time comparisons

+++ system/None/thread.cpp
@@ -10,7 +10,7 @@ Thread Thread::Current(void) { return Thread(); }
 void Thread::Sleep(uint32_t milliseconds)
 {
   uint32_t now = Timebase::GetMilliseconds();
-  while (Timebase::GetMilliseconds() - now < milliseconds)
+  while (uint32_t(Timebase::GetMilliseconds()) - now < milliseconds)
   {
     Timer::RefreshTimerInIdle();
   }
@@ -18,7 +18,7 @@ void Thread::Sleep(uint32_t milliseconds)
 
 void Thread::SleepUntil(TimestampMS &last_waskup_time, uint32_t time_to_sleep)
 {
-  while (Timebase::GetMilliseconds() - last_waskup_time < time_to_sleep)
+  while (uint32_t(Timebase::GetMilliseconds()) - last_waskup_time < time_to_sleep)
   {
     Timer::RefreshTimerInIdle();
   }
